requirements.txt
# LIFEE 依赖

# 核心
anthropic>=0.40.0      # Claude API
pydantic>=2.0          # 数据验证
pydantic-settings>=2.0 # 配置管理
python-dotenv>=1.0.0   # 环境变量

# LLM Providers
openai>=1.0.0          # OpenAI 兼容 API (Qwen, Ollama, OpenCode)
google-genai>=1.0.0    # Google Gemini API

# 工具
httpx>=0.25.0          # HTTP 客户端
questionary>=2.0.0     # 交互式命令行选择

# 知识库/RAG
numpy>=1.24.0          # 向量计算（余弦相似度）

# 市场数据
yfinance>=0.2.36       # 实时股票数据（美股/港股/A股）

# 工具能力
ddgs>=7.0.0            # DuckDuckGo 搜索（角色联网）
